Refactor membership functions out from lib.user
This is needed because the traffic functions depend on them, causing
cyclic imports if we would introduce them in a separate module like
`lib.traffic`.  With this restructuring, it is safely possible.
